# Better batching rule for triangular_solve