--- a/jax/lax_linalg.py
+++ b/jax/lax_linalg.py
@@ -318,6 +318,9 @@ def triangular_solve_shape_rule(a, b, left_side=False, **unused_kwargs):
   if a.ndim < 2:
     msg = "triangular_solve requires a.ndim to be at least 2, got {}."
     raise TypeError(msg.format(a.ndim))
+  if b.ndim < 2:
+    msg = "triangular_solve requires b.ndim to be at least 2, got {}."
+    raise TypeError(msg.format(b.ndim))
   if a.shape[-1] != a.shape[-2]:
     msg = ("triangular_solve requires the last two dimensions of a to be equal "
            "in size, got a.shape of {}.")