Now, if only the right hand side argument `b` is batched, we leverage
triangular solve's builtin batching for handling multiple right-hand-side
vectors.

This makes the performance of `vmap` over only the second argument of linear
solves equivalent to relying on builtin batching::

    rs = onp.random.RandomState(0)
    a = rs.randn(500, 500) + 0.1 * np.eye(500)
    b_mat = jax.device_put(rs.randn(500, 10))
    solve1 = jax.jit(np.linalg.solve)
    solve2 = jax.jit(jax.vmap(np.linalg.solve, in_axes=(None, 1), out_axes=1))

Before::

    In [6]: %timeit jax.device_get(solve1(a, b_mat))
    3.88 ms ± 293 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

    # 8x slower :(
    In [9]: %timeit jax.device_get(solve2(a, b_mat))
    23.5 ms ± 1.33 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

Now::

    In [2]: %timeit jax.device_get(solve1(a, b_mat))
    3.76 ms ± 304 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)

    # same speed :)
    In [3]: %timeit jax.device_get(solve2(a, b_mat))
    3.72 ms ± 296 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)